Move config traits to core so we can use in chains
