+++ rust/hyperlane-base/src/settings/base.rs
@@ -6,20 +6,21 @@ use futures_util::StreamExt;
 use itertools::Itertools;
 use serde::Deserialize;
 
-use hyperlane_core::utils::StrOrInt;
 use hyperlane_core::{
+    config::*,
     db::{HyperlaneDB, DB},
     HyperlaneChain, HyperlaneDomain, HyperlaneProvider, InterchainGasPaymaster,
     InterchainGasPaymasterIndexer, Mailbox, MailboxIndexer, MultisigIsm, ValidatorAnnounce, H256,
 };
 
-use crate::chains::RawChainConf;
 use crate::{
-    settings::{chains::ChainConf, signers::SignerConf, trace::TracingConfig},
-    CachingInterchainGasPaymaster, ConfigErrResultExt, ConfigParsingError, ConfigPath,
-    ConfigResultExt, FromRawConf, IntoParsedConf, RawSignerConf,
+    settings::{
+        chains::{ChainConf, RawChainConf},
+        signers::SignerConf,
+        trace::TracingConfig,
+    },
+    CachingInterchainGasPaymaster, CachingMailbox, CoreMetrics, HyperlaneAgentCore, RawSignerConf,
 };
-use crate::{CachingMailbox, CoreMetrics, HyperlaneAgentCore};
 
 /// Settings. Usually this should be treated as a base config and used as
 /// follows:
