+++ rust/hyperlane-base/src/settings/mod.rs
@@ -69,23 +69,11 @@
 //!    intended to be used by a specific agent.
 //!    E.g. `export HYP_RELAYER_ORIGINCHAIN="ethereum"`
 
-use std::fmt::{Debug, Display, Formatter};
-use std::ops::Deref;
-use std::rc::Rc;
-
-use eyre::{eyre, Context, Report};
-use futures_util::AsyncReadExt;
-use itertools::Itertools;
 use once_cell::sync::OnceCell;
 use rusoto_kms::KmsClient;
-use serde::Deserialize;
 
 pub use base::*;
 pub use chains::{ChainConf, ChainConnectionConf, CoreContractAddresses};
-use hyperlane_core::{
-    HyperlaneChain, HyperlaneProvider, InterchainGasPaymaster, InterchainGasPaymasterIndexer,
-    Mailbox, MailboxIndexer, MultisigIsm, ValidatorAnnounce,
-};
 pub use signers::{RawSignerConf, SignerConf};
 
 use crate::CachingInterchainGasPaymaster;
@@ -101,223 +89,3 @@ mod signers;
 pub mod trace;
 
 static KMS_CLIENT: OnceCell<KmsClient> = OnceCell::new();
-
-// TODO: Remove extension functions we are not using
-
-pub trait ConfigOptionExt<T> {
-    fn expect_or_eyre<M: Into<String>>(self, msg: M) -> eyre::Result<T>;
-
-    fn expect_or_else_eyre<S, F>(self, f: F) -> eyre::Result<T>
-    where
-        S: Into<String>,
-        F: FnOnce() -> S;
-
-    fn expect_or_parsing_error<S, F>(self, v: F) -> ConfigResult<T>
-    where
-        S: Into<String>,
-        F: FnOnce() -> (ConfigPath, S);
-}
-
-impl<T> ConfigOptionExt<T> for Option<T> {
-    fn expect_or_eyre<M: Into<String>>(self, msg: M) -> eyre::Result<T> {
-        self.ok_or_else(|| eyre!(msg.into()))
-    }
-
-    fn expect_or_else_eyre<S, F>(self, f: F) -> eyre::Result<T>
-    where
-        S: Into<String>,
-        F: FnOnce() -> S,
-    {
-        self.ok_or_else(|| eyre!(f().into()))
-    }
-
-    fn expect_or_parsing_error<S, F>(self, v: F) -> ConfigResult<T>
-    where
-        S: Into<String>,
-        F: FnOnce() -> (ConfigPath, S),
-    {
-        self.ok_or_else(|| {
-            let (path, msg) = v();
-            ConfigParsingError::new(path, eyre!(msg.into()))
-        })
-    }
-}
-
-pub trait ConfigErrResultExt<T> {
-    fn into_config_result(self, path: impl FnOnce() -> ConfigPath) -> ConfigResult<T>;
-
-    fn merge_err_then_none(
-        self,
-        err: &mut ConfigParsingError,
-        path: impl FnOnce() -> ConfigPath,
-    ) -> Option<T>;
-
-    fn merge_err_with_ctx_then_none<S, F>(self, err: &mut ConfigParsingError, ctx: F) -> Option<T>
-    where
-        S: Into<String>,
-        F: FnOnce() -> (ConfigPath, S);
-}
-
-impl<T, E> ConfigErrResultExt<T> for Result<T, E>
-where
-    E: Into<Report>,
-{
-    fn into_config_result(self, path: impl FnOnce() -> ConfigPath) -> ConfigResult<T> {
-        self.map_err(|e| ConfigParsingError::new(path(), e.into()))
-    }
-
-    fn merge_err_then_none(
-        self,
-        err: &mut ConfigParsingError,
-        path: impl FnOnce() -> ConfigPath,
-    ) -> Option<T> {
-        match self {
-            Ok(v) => Some(v),
-            Err(e) => {
-                err.merge(ConfigParsingError::new(path(), e));
-                None
-            }
-        }
-    }
-
-    fn merge_err_with_ctx_then_none<S, F>(self, err: &mut ConfigParsingError, ctx: F) -> Option<T>
-    where
-        S: Into<String>,
-        F: FnOnce() -> (ConfigPath, S),
-    {
-        match self {
-            Ok(v) => Some(v),
-            Err(e) => {
-                let (path, ctx) = ctx();
-                let report: Report = Err::<(), _>(e.into()).context(ctx.into()).err().unwrap();
-                err.merge(ConfigParsingError::new(path, report));
-                None
-            }
-        }
-    }
-}
-
-pub trait ConfigResultExt<T> {
-    fn merge_parsing_err_then_none(self, err: &mut ConfigParsingError) -> Option<T>;
-}
-
-impl<T> ConfigResultExt<T> for ConfigResult<T> {
-    fn merge_parsing_err_then_none(self, err: &mut ConfigParsingError) -> Option<T> {
-        match self {
-            Ok(v) => Some(v),
-            Err(e) => {
-                err.merge(e);
-                None
-            }
-        }
-    }
-}
-
-// declare_deserialize_for_config_struct!(Settings);
-
-#[derive(Debug, Default, PartialEq, Eq, Clone)]
-pub struct ConfigPath(Vec<Rc<String>>);
-
-impl Display for ConfigPath {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.path_name())
-    }
-}
-
-impl ConfigPath {
-    pub fn join(&self, part: impl Into<String>) -> Self {
-        let part = part.into();
-        debug_assert!(!part.contains('.'));
-        let mut new = self.clone();
-        new.0.push(Rc::new(part));
-        new
-    }
-
-    pub fn merge(&self, other: &Self) -> Self {
-        Self(
-            self.0
-                .iter()
-                .cloned()
-                .chain(other.0.iter().cloned())
-                .collect(),
-        )
-    }
-
-    pub fn path_name(&self) -> String {
-        self.0.iter().map(|s| s.as_str()).join(".")
-    }
-
-    pub fn env_name(&self) -> String {
-        ["HYP", "BASE"]
-            .into_iter()
-            .chain(self.0.iter().map(|s| s.as_str()))
-            .map(|s| s.to_uppercase())
-            .join("_")
-    }
-}
-
-#[derive(Debug, Default)]
-pub struct ConfigParsingError(Vec<(ConfigPath, Report)>);
-
-impl ConfigParsingError {
-    pub fn new(path: ConfigPath, report: impl Into<Report>) -> Self {
-        Self(vec![(path, report.into())])
-    }
-
-    pub fn from_report(conf_path: ConfigPath, report: Report) -> Self {
-        Self(vec![(conf_path, report)])
-    }
-
-    pub fn push(&mut self, conf_path: ConfigPath, report: Report) {
-        self.0.push((conf_path, report));
-    }
-
-    pub fn merge(&mut self, other: Self) {
-        self.0.extend(other.0);
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-}
-
-pub type ConfigResult<T> = Result<T, ConfigParsingError>;
-
-impl FromIterator<ConfigParsingError> for ConfigParsingError {
-    fn from_iter<T: IntoIterator<Item = ConfigParsingError>>(iter: T) -> Self {
-        Self(iter.into_iter().flat_map(|e| e.0).collect())
-    }
-}
-
-impl Display for ConfigParsingError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "ParsingError [")?;
-        for (path, report) in &self.0 {
-            write!(f, "({path}: {report}),")?;
-        }
-        write!(f, "]")
-    }
-}
-
-impl std::error::Error for ConfigParsingError {}
-
-pub trait FromRawConf<'de, T>: Sized
-where
-    T: Debug + Deserialize<'de>,
-{
-    fn from_config(raw: T, cwp: &ConfigPath) -> ConfigResult<Self>;
-}
-
-pub trait IntoParsedConf<'de, O: Sized>: Debug + Deserialize<'de> {
-    fn parse_config(self, cwp: &ConfigPath) -> ConfigResult<O>;
-}
-
-impl<'de, S, O> IntoParsedConf<'de, O> for S
-where
-    S: Deserialize<'de> + Debug,
-    O: FromRawConf<'de, S>,
-{
-    fn parse_config(self, cwp: &ConfigPath) -> ConfigResult<O> {
-        O::from_config(self, cwp)
-    }
-}
