+++ rust/hyperlane-core/Cargo.toml
@@ -18,7 +18,9 @@ derive-new.workspace = true
 ethers-providers.workspace = true
 ethers-signers.workspace = true
 ethers.workspace = true
+eyre.workspace = true
 hex = "0.4.3"
+itertools.workspace = true
 lazy_static = "*"
 num = {version="0", features=["serde"]}
 num-derive = "0.3"
