+++ rust/hyperlane-core/src/utils.rs
@@ -1,11 +1,5 @@
-use std::fmt::{Debug, Formatter};
-use std::num::{ParseIntError, TryFromIntError};
-
-use serde::Deserialize;
+use crate::{KnownHyperlaneDomain, H256};
 use sha3::{digest::Update, Digest, Keccak256};
-use thiserror::Error;
-
-use crate::{KnownHyperlaneDomain, H256, U256};
 
 /// Strips the '0x' prefix off of hex string so it can be deserialized.
 ///
@@ -61,107 +55,6 @@ pub fn fmt_domain(domain: u32) -> String {
     KnownHyperlaneDomain::try_from(domain)
         .map(|d| d.to_string())
         .unwrap_or_else(|_| domain.to_string())
-}
-
-/// An error when parsing a StrOrInt type as an integer value.
-#[derive(Error, Debug)]
-pub enum StrOrIntParseError {
-    /// The string is not a valid integer
-    #[error("Invalid integer provided as a string: {0}")]
-    StrParse(#[from] ParseIntError),
-    /// The provided integer does not match the type requirements.
-    #[error("Provided number is an invalid integer: {0}")]
-    InvalidInt(#[from] TryFromIntError),
-    #[error("Could not parse integer: {0}")]
-    Other(String),
-}
-
-/// A type which can be used for parsing configs that may be provided as a
-/// string or an integer but will ultimately be read as an integer. E.g. where
-/// `"domain": "42"` and `"domain": 42` should both be considered valid.
-#[derive(Clone, Deserialize)]
-#[serde(untagged)]
-pub enum StrOrInt {
-    /// The parsed type is a string
-    Str(String),
-    /// The parsed type is an integer
-    Int(i64),
-}
-
-impl Debug for StrOrInt {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            StrOrInt::Str(v) => write!(f, "\"{v}\""),
-            StrOrInt::Int(v) => write!(f, "{}", *v),
-        }
-    }
-}
-
-impl From<i64> for StrOrInt {
-    fn from(value: i64) -> Self {
-        StrOrInt::Int(value)
-    }
-}
-
-impl From<String> for StrOrInt {
-    fn from(value: String) -> Self {
-        StrOrInt::Str(value)
-    }
-}
-
-impl From<&str> for StrOrInt {
-    fn from(value: &str) -> Self {
-        StrOrInt::Str(value.to_owned())
-    }
-}
-
-macro_rules! convert_to {
-    ($t:ty) => {
-        impl TryFrom<StrOrInt> for $t {
-            type Error = StrOrIntParseError;
-
-            fn try_from(v: StrOrInt) -> Result<Self, Self::Error> {
-                (&v).try_into()
-            }
-        }
-
-        impl TryFrom<&StrOrInt> for $t {
-            type Error = StrOrIntParseError;
-
-            fn try_from(v: &StrOrInt) -> Result<Self, Self::Error> {
-                Ok(match v {
-                    StrOrInt::Str(s) => s.parse()?,
-                    StrOrInt::Int(i) => (*i).try_into()?,
-                })
-            }
-        }
-    };
-}
-
-convert_to!(u16);
-convert_to!(u32);
-convert_to!(u64);
-
-impl TryFrom<StrOrInt> for U256 {
-    type Error = StrOrIntParseError;
-
-    fn try_from(v: StrOrInt) -> Result<Self, Self::Error> {
-        (&v).try_into()
-    }
-}
-impl TryFrom<&StrOrInt> for U256 {
-    type Error = StrOrIntParseError;
-
-    fn try_from(v: &StrOrInt) -> Result<Self, Self::Error> {
-        Ok(match v {
-            StrOrInt::Str(s) => s.parse().map_err(|_| {
-                StrOrIntParseError::Other(format!("Unable to parse U256 string ({s})"))
-            })?,
-            StrOrInt::Int(i) => (*i).try_into().map_err(|_| {
-                StrOrIntParseError::Other(format!("Unable to parse integer as U256 ({i})"))
-            })?,
-        })
-    }
 }
 
 /// Shortcut for many-to-one match statements that get very redundant. Flips the
